Remove ensureCacheLimit method and calls from buildSrc. (#66480)
The ensureCacheLimit method would check to see if the uncompressed image
exceeded 400MB, and fail the build if it exceeded this size. This was
implemented to have a means of determining when a docker image size
would exceed the allowed cachable size of an artifact by our Gradle
Enterprise server. This is important to us because when the size exceeds
the 400MB limit, the container is rebuilt every time rather than
downloaded from the remote cache. However, Gradle will compress this
image before sending the artifact to the remote server, which this did
not properly account for.

As result, we are removing this method for now to unblock ongoing work
until we implement a check which is able to determine the actual size of
the compressed artifact.