Migrate python-services-v3 into Gradle build. (#76004)
* Migrate docker image from Dockerfile to Gradle for python-services-v3.

This creates the docker images for python-services-v3 in Gradle. There
are some key changes in how we produce the compose environment. One major
change is the generation of the certificates is no longer done within a
volume. Instead, we produce an intermediate container and export the
resulting files which are bind mounted into the containers which use them.

* Remove python-services-v3 build stage from Jenkins.

The build is now executed as part of the Gradle build. As result, this
build stage is no longer required.

* Fix broken report generation from xmlrunner/tox.

The xmlrunner output would overwrite the junit.xml between unit and
integration tests. This prevented full test reports from being pulled
into Gradle Enterprise.

* Remove deprecated Dockerfiles.

* Disable coverage reports for python-services-v3 in Gradle.

The coverage reports are not fully functioning within the Gradle
context. Moreover, we have not yet added support to ingest them into
Gradle Enterprise as of yet. This disables the coverage reports for
now until we can fix the issue and expose them in Gradle Enterprise.

* Add :python-services-v3:services-extra Gradle task.

This migrates the start/stop services-extra from the Makefile into
Gradle.

* Remove python-services-v3 Makefile and docker-compose.docs.yml

The core features and functionality provided by the Makefile has
been migrated to the Gradle build. In addition, the `serve-docs`
make target is not needed or used and as a result I am removing
it from the project.


* Migrate python-services-app from makefile to Gradle.

* DRY out Stack Version used in compose files and Gradle files.

This creates a single variable to set the stack version across all
the containers in the python-services-v3 build.

* Use dynamic host port allocation for postgres container.

When running the Gradle build on CI workers we have a high level of
parallelism. In some cases, the postgresql port is in use by another
instance of postgreqsl in use by a different portion of the build.
This changes docker compose to use a dynamic host port allocation to
ensure that we do not run into issues starting the postgres container
for python-services-v3 testing.

* Build billing-service, python-services-{audit,cron} in Gradle

This updates the build and push of python-service-v3 docker images
billing-service, python-services-cron, and python-services-audit to
use Gradle instead of make.

* Move tox XML reports into buildDir.

This moves the tox XML reports into the buildDir to ensure they are
removed between each build and so that Git doesn't see them as
trackable files.

* Set package versions and create base image to pre-install gnupg.

The Docker plugin attempts to collapse commands into a single layer
and execution step. In this case, it would cause it to attempt to
configure the deadsnakes Python PPA and GPG keys before it installs
the gnupg package. The only workaround is to create a base-image
with the gnupg package pre-installed before we attempt to configure
the image. This is required to pin the version of Python we wish
to install.


* Disable scala test class which depends on removed Dockerfile.

See issue: https://github.com/elastic/cloud/issues/78190

* Increase nofile ulimit for python-services test containers.

We have been seeing test failures when running the python-services
tests. This is an attempt to try and mitigate the failures.

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
Co-authored-by: Alpar Torok <alpar.torok@elastic.co>