Component image multi platform support (#82884)
## Notes for your reviewers
Adds the DSL to be able to express multi platform component images. 

Note that this PR is for the DSL only, it uses the locally built base image as a base for both. 
Also the result is two separate images, not combined in a manifest list yet. 
The build info stuff is not ideal  and confusing, but since we'll clean it up in #82551 I haven't payed much attention to it. 


This is how the DSL looks like 
```  
       from(project(":base:image:project")
       images(
            mapOf(
                OS.LINUX to Architecture.X86_64
            )
        ) {
             if (architecture == Architecutre.AARCH64) {
                // specific setup for ARM
            }
            copySpec {  from("proxy-$os_$architecture")  }
        }
```
It moves the layer definitions inside an `images() {}` block that also defines the platforms we want to build for.
The `os` and `architecture` variables can be reffered to to decide on what goes into the layers of each image.
That of course means the DSL is evaluated once for every platrorm.     

The push task doesn't add the platform to `linux_x86_64`  so everything will work as it should untill we add the manifest list. 

## What issues does this PR fix?
Part of: #75776