Convert python services to the new base/component plugins (#83575)
Notes for your reviewers

While converting images, some changes to the plugins were needed:

    component images were missing a working dir
    both base and component images were being skipped in case there was no copySpec instructions because of how the underlying Sync tasks, so we always create an empty .elastic file in the build dir to be copied in the root of the context dir to make sure the task is never skipped.
    docker sandbox exec task gets support for the new base plugin
    The DB migrations image no longer uses the S3 bucket, the artifact is now in artifactory

Most changes just convert the build script to the equivalent base and component images. The one to note is that found-services now has it's own project. This is required because there are no push aliases in the new tooling I think this creates a much better mapping between Gradle projects and docker images and makes it easier to track a docker image back to source.
What issues does this PR fix?

Part of #82553