Filter out common system binaries from inputs for SandboxExecTask (#91651)
We had extra builds happening, because of cache misses, from common system binaries differing slightly across platforms. Implementations of things like `cp` might differ at the byte-level, but are so ubiquitous that we can trust their implementation across platforms. This code adds in a small whitelist to allow these very common binaries to be filtered out from the inputs, resulting in more cache hits and faster build times. A mechanism to override the list of binaries was added, just in case.

Closes https://github.com/elastic/cloud/issues/79462