Fix failing https://buildkite.com/elastic/gradle-plugins-ci builds.
They were failing in co.elastic.gradle.dockerbase.DockerBaseImageMatrix for the Wolfi base images with an error and
`Could not find wolfi/aarch64:__META__Packages` and `Could not find wolfi/x86_64:__META__Packages`

- Sample failure: https://buildkite.com/elastic/gradle-plugins-ci/builds/265
- Test failure link 1: https://gradle-enterprise.elastic.co/s/nlhfbci6vfyxu/tests/task/:plugins:docker:base-image:integrationTest/details/co.elastic.gradle.dockerbase.DockerBaseImageMatrix/testSingleProject(String%2C%20Path)%5B1%5D?top-execution=1
- Test failure link 2: https://gradle-enterprise.elastic.co/s/3ctsxah6nkuv2/tests/task/:plugins:docker:base-image:integrationTest/details/co.elastic.gradle.dockerbase.DockerBaseImageMatrix/testSingleProject(String%2C%20Path)%5B1%5D?top-execution=1

Looking at e.g. https://github.com/adoptium/installer/issues/689 (for an unrelated project, but indicating a similar
issue) it seems that the Alpine APK tool is unable to verify package signatures for local packages. Specifying
`--allow-untrusted` to bypass package verification when creating the local package index fixes the gradle-plugins-ci
build issues.

Since the packages were downloaded and installed just prior to the `apk index` command, I think the `allow-untrusted`
flag here is not much of a risk.

It seems that at some point, either the Wolfi packages stopped being signed and made `apk index` start failing as it
expects to verify file signatures when doing so or, as the above GitHub issue indicates, package verification for local
packages as opposed to downloaded from a (signed) repository is just different.

Comparing an `apk-tools` package downloaded from Alpine Linux as opposed to Wolfi shows that the latter is missing an
embedded public key that can be used to perform the necessary validation, perhaps supporting the former scenario:

```
$ curl -s https://dl-cdn.alpinelinux.org/alpine/v3.18/main/aarch64/apk-tools-2.14.4-r0.apk | tar ztf - | grep '^\.'
.SIGN.RSA.alpine-devel@lists.alpinelinux.org-616ae350.rsa.pub
.PKGINFO

$ curl -s https://packages.wolfi.dev/os/aarch64/apk-tools-2.14.4-r1.apk | tar ztf - | grep '^\.'
.PKGINFO
```

Either way, I would propose we make this change as the gradle-plugins-ci build has been failing for a long time now and
is holding up merges to `main`.