feat(vault): allow specifying `VAULT_AUTH_PATH` for AppRole
In the case that we're authenticating via an AppRole, but aren't using
the default `approle` auth path, we want to allow wiring in via a new
environment variable, `VAULT_AUTH_PATH`.

This duplicates the core integration test from
`canReadVaultSecretsWithRoles`, but with our new Vault auth path of
`new_path`.

If not set, this will default to `approle`.