fix(conventions): correctly set a convention for Vault address
In some cases, we're seeing that a Vault address explicitly set in the
buildscript i.e.

```kotlin
vault {
  address.set("https://...")
}
```

will not be correctly wired in, and instead the default value of
`vault-ci-prod.elastic.dev` is used.

This appears to be happening intermittently, which is why we've not seen
it up until now.

It appears that this may be due to the usage of `.set(...)` which
overrides any existing value, instead of defining the convention (aka
default).